django/requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.0
cssselect==1.1.0
decorator==4.4.0
Django==2.2.4
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
lxml==4.4.1
markdown2==2.3.8
numpy==1.17.0
pandas==0.24.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pypandoc==1.4
pyquery==1.4.0
python-dateutil==2.8.0
python-slugify==3.0.3
pytz==2019.2
readtime==1.1.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
